connect logic to separate hook
