+++ src/components/_app/Web3ConnectionManager/Web3ConnectionManager.tsx
@@ -5,14 +5,9 @@ import { WalletConnect } from "@web3-react/walletconnect"
 import AccountModal from "components/common/Layout/components/Account/components/AccountModal"
 import NetworkModal from "components/common/Layout/components/Account/components/NetworkModal/NetworkModal"
 import requestNetworkChangeHandler from "components/common/Layout/components/Account/components/NetworkModal/utils/requestNetworkChange"
-import useUser from "components/[guild]/hooks/useUser"
-import { useConnect } from "components/[guild]/JoinModal/hooks/useConnectPlatform"
-import { Message } from "components/[guild]/JoinModal/hooks/useOauthPopupWindow"
 import { Chains, RPC } from "connectors"
-import useKeyPair from "hooks/useKeyPair"
 import useToast from "hooks/useToast"
 import { useRouter } from "next/router"
-import platforms from "platforms/platforms"
 import {
   createContext,
   PropsWithChildren,
@@ -20,8 +15,8 @@ import {
   useEffect,
   useState,
 } from "react"
-import useDatadog from "../Datadog/useDatadog"
 import WalletSelectorModal from "./components/WalletSelectorModal"
+import useConnectFromLocalstorage from "./hooks/useConnectFromLocalStorage"
 import useEagerConnect from "./hooks/useEagerConnect"
 
 const Web3Connection = createContext({
@@ -48,47 +43,10 @@ const Web3Connection = createContext({
 const Web3ConnectionManager = ({
   children,
 }: PropsWithChildren<any>): JSX.Element => {
-  const { keyPair, isValid } = useKeyPair()
-  const { addDatadogError } = useDatadog()
-  const toast = useToast()
-  const { onSubmit } = useConnect(() => {
-    toast({ status: "success", title: "Success", description: "Platform connected" })
-  })
-  const { platformUsers } = useUser()
-
-  useEffect(() => {
-    if (!keyPair || !isValid || !platformUsers) return
-
-    Object.keys(platforms).forEach((platformName) => {
-      const storageKey = `${platformName}_shouldConnect`
-      const strData = window.localStorage.getItem(storageKey)
-      window.localStorage.removeItem(storageKey)
-
-      const isAlreadyConnected = platformUsers.some(
-        (platformUser) => platformUser.platformName === platformName
-      )
-      if (isAlreadyConnected) return
-
-      if (strData) {
-        const data: Message = JSON.parse(strData)
-
-        if (data.type === "OAUTH_SUCCESS") {
-          onSubmit({ platformName, authData: data.data })
-        } else {
-          toast({
-            status: "error",
-            title: data.data.error ?? "Error",
-            description:
-              data.data.errorDescription || `Failed to connect ${platformName}`,
-          })
-          addDatadogError("OAuth error from localStorage data", data.data)
-        }
-      }
-    })
-  }, [keyPair, isValid, platformUsers])
-
   const { isActive, connector } = useWeb3React()
   const router = useRouter()
+
+  useConnectFromLocalstorage()
 
   const {
     isOpen: isWalletSelectorModalOpen,
@@ -117,6 +75,7 @@ const Web3ConnectionManager = ({
   }, [triedEager, isActive, router.query])
 
   const [isNetworkChangeInProgress, setNetworkChangeInProgress] = useState(false)
+  const toast = useToast()
   const requestManualNetworkChange = (chain) => () =>
     toast({
       title: "Your wallet doesn't support switching chains automatically",
